+++ CHANGELOG.md
@@ -21,3 +21,5 @@ Sections should be in the alphabetic order for each release:
 - [3. Gem Installation and Configuration](https://trello.com/c/CMYbiv9J/2-3-gem-installation-and-configuration)
 - [4. Database Setup](https://trello.com/c/0jMOW919/4-4-database-setup)
 - [16. Implement User Model for API Authentication](https://trello.com/c/8hMt0TyS/16-16-implement-user-model-for-api-authentication)
+- [5. Create Task Model and Migration](https://trello.com/c/ucmsH1zD/5-5-create-task-model-and-migration)
+- [8. Task Validations](https://trello.com/c/ovKAfU6u/8-8-task-validations)